Add files via upload
Just a set of super base classes from java that I made while self-teaching with no internet access. Useful or not, I'm a n00b and don't know, but thought it was a good exercise at the time(consider is a game board a thing or a place...). Can they be better? Definitely. I would like Thing and Person to conditionally extend Place given a range or position is true, but that is thus far beyond my knowledge.
*re because my initials, and re-usable.
*yes, my commenting sucks.
*Do I need a disclaimer, really? standard open source rules or whatever.